lesson_1_2 assignment successfully completed
